requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
rdkit-pypi==2022.9.5
